refactor(document): 重构异步处理逻辑实现
- 添加了多个字符串处理和切片操作相关的导入包
- 移除了硬编码的token估算系数常量
- 重写了HandleIndexBuild方法，增加了事务处理和异常恢复机制
- 优化了文档索引构建的任务状态管理和日志记录流程
- 改进了父块和子块实体构建中的字符计数方式
- 增强了错误处理和任务失败恢复的逻辑实现

+++ internal/domain/document/model/entity/document_chunk.go
@@ -14,7 +14,7 @@ type DocumentChunk struct {
 	StructureNodeId    int64  `gorm:"column:structure_node_id"`   // 结构节点ID
 	StructureNodeType  int    `gorm:"column:structure_node_type"` // 结构节点类型
 	CanonicalPath      string `gorm:"column:canonical_path"`      // 规范路径
-	ItemIndex          int64  `gorm:"column:item_index"`          // 项索引
+	ItemIndex          int    `gorm:"column:item_index"`          // 项索引
 	ChunkText          string `gorm:"column:chunk_text"`          // 块文本
 	CharCount          int    `gorm:"column:char_count"`          // 字符数
 	TokenCount         int    `gorm:"column:token_count"`         // token数
